back to image_copier list strategy
