+++ Cargo.toml
@@ -19,6 +19,7 @@ ratatui = "0.29.0"
 ratatui-image = "3.0.0"
 bevy_ratatui = "0.7.0"
 image = "0.25.5"
+bevy_mod_debugdump = "0.12.0"
 
 [profile.dev]
 opt-level = 1
