--- src/assets.rs
@@ -1,7 +1,0 @@
-use bevy::asset::embedded_asset;
-use bevy::prelude::*;
-
-pub(super) fn plugin(app: &mut App) {
-    let prefix = "src/";
-    embedded_asset!(app, prefix, "shaders/sobel.wgsl");
-}
